+++ src/dags/pipelinedefinition.cpp
@@ -394,7 +394,7 @@ public:
         SPDLOG_LOGGER_DEBUG(modelmanager_logger, "Validation of pipeline: {}; node name: {}; node kind: {}",
             pipelineName,
             dependantNodeInfo.nodeName,
-            (unsigned int)dependantNodeInfo.kind);
+            dependantNodeInfo.kind);
     }
 
     Status fetchUnderlyingModelInstance() {
